move version from pkg to internal
